requirements.txt
pandas
openpyxl
python-dotenv
python-pptx

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
